chore: Separate interface from implementation for Variant
